pyproject.toml
[project]
name = "youtube-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.115.0",
    "groq==0.9.0",
    "httpx==0.27.2",
    "pydantic==2.8.2",
    "pydantic-settings==2.4.0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.9",
    "yt-dlp==2026.3.17",
    "uvicorn[standard]==0.30.6",
    "youtube-transcript-api==1.2.4",
]
